+++ CHANGELOG.md
@@ -17,6 +17,10 @@ Given the `MAJOR.MINOR.PATCH` pattern, here is how we decide to increment:
 
 ### Added
 
+- Added tests for the public methods in the Multiselect
+- API client for Regulations.gov Comment API
+- reg-comment organism and Wagtail module for adding it to a page
+
 ### Changed
 
 ### Removed